requirements.txt
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
pdfplumber==0.11.6
python-dotenv==1.1.0
streamlit==1.44.1
openai==1.76.0
faiss-cpu==1.11.0
altair==5.5.0
tiktoken==0.9.0